+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Features
 - UI: The first item in a combobox will automatically be submitted if enter is pressed without a selection.
+- Added a search function to the property editor (advanced mode)
+- Added a new (old) property type: `byteArray`. Their `Value` is always the same as their `Bytes`.
 
 ## Bugfixes
 - Fixed the "freeze" button turning completely blue, instead of just the icon inside.